third_party/WebKit/LayoutTests/inspector/tracing/timeline-script-tag-1.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>

function performActions()
{
    var iframe = document.createElement("iframe");
    iframe.src = "resources/timeline-iframe-data.html";
    document.body.appendChild(iframe);
}

function test() 
{
    WebInspector.panels.timeline._captureJSProfileSetting.set(false);
    InspectorTest.startTimeline(step1);
    function step1()
    {
        InspectorTest.addConsoleSniffer(step2);
        InspectorTest.evaluateInPage("performActions()");
    }

    function step2()
    {
        InspectorTest.stopTimeline(step3);
    }

    function step3()
    {
        function format(record)
        {
            if (record.type() === WebInspector.TimelineModel.RecordType.EvaluateScript) {
                InspectorTest.dumpTimelineRecord(record, undefined, undefined, [
                    "TimeStamp",
                ]);
                InspectorTest.printTimelineRecordProperties(record);
            }
        }

        InspectorTest.printTimelineRecords(null, InspectorTest.safeWrap(format));
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline API instrumentation of an HTML script tag.
</p>

</body>
</html>
